// backend/services/devops-agent/src/utils/logger.d.ts
export declare const logger: any;
export declare const createContextLogger: (context: string) => {
    error: (message: string, ...args: any[]) => any;
    warn: (message: string, ...args: any[]) => any;
    info: (message: string, ...args: any[]) => any;
    http: (message: string, ...args: any[]) => any;
    debug: (message: string, ...args: any[]) => any;
};
export default logger;
//# sourceMappingURL=logger.d.ts.map